refactor(sofast-module-demo):重构 bean search 示例
- 更新 OrderReq 和 OrderRes 类，添加必要的注解和继承
- 重构 BeanSearchController，使用 SearchBeanUtils 进行查询
- 移除 MyParamFilter，改为在框架中统一处理逻辑删除
- 更新 PageParam 类，为 Swagger 文档隐藏部分字段
- 新增 LogicDeleteParamFilter 和 SearchBeanBase 类
- 添加 SearchBeanUtils 工具类，简化查询操作
- 更新 PageUtil 类，统一分页结果处理
- 在 SofastMybatisAutoConfiguration 中注册 LogicDeleteParamFilter

+++ sofast-spring-boot-starter/sofast-spring-boot-starter-mybatis/src/main/java/cc/sofast/framework/starter/mybatis/utils/PageUtil.java
@@ -37,15 +37,20 @@ public class PageUtil {
         return pageResult;
     }
 
-
-    public static <T> PageResult<T> toPageResult(SearchResult<T> search,Long pageNum) {
+    /**
+     * 将mybatis-plus分页结果转换为通用分页结果
+     *
+     * @param search mybatis-plus分页结果
+     * @return PageResult<T>
+     */
+    public static <T> PageResult<T> toPageResult(SearchResult<T> search, int pageNum) {
         if (search == null) {
             return new PageResult<>();
         }
         PageResult<T> pageResult = new PageResult<>();
         List<T> dataList = search.getDataList();
         pageResult.setPageSize((long) dataList.size());
-        pageResult.setPageIndex(pageNum);
+        pageResult.setPageIndex((long) pageNum);
         pageResult.setTotalCount(search.getTotalCount().longValue());
         pageResult.setData(search.getDataList());
         return pageResult;
